Collision detection using getClientRects() instead of bound client rectangle.
This solves a tricky problem whereby the union of getClientRects was overlapping. Instead, we now treat each selection as a separate claimed rectangle.

This creates another difficulty in ensuring that cross-browser behaviour is consistent. Indeed, in Chrome it was turning out that a bounding box was added, for every line except the first, that was precisely 12px less than the viewport.